0.8.1 Bug Corrections
Current task involves correcting errors in the minimap. This push for 0.8.1 corrects an issue with the zoom function, text scaling, and map centering.  Coin Scraper class has been removed and that functionality has been added in to a real-time html scrape. Zoom levels limited to 3x3,5x5 and 7x7 for text label effectiveness. All minimap updates are now moved to draw_minimap which are called through a request to update_minimap.